Fix initial setup of units from env var
Initial units were previously setup in the start.sh script but
this would create a race where units could not be created fast enough
before the tests in the sut container ran. This moves the unit
creation to main.rs so the units can be made available as soon as the
server is setup.

Change-type: patch